+++ internal/write/service.go
@@ -57,25 +57,25 @@ func NewService(org string, bucket string, httpService ihttp.Service, options *w
 }
 
 func (w *Service) HandleWrite(ctx context.Context, batch *Batch) error {
-	log.Log.Debug("Write proc: received write request")
+	log.Debug("Write proc: received write request")
 	batchToWrite := batch
 	retrying := false
 	for {
 		select {
 		case <-ctx.Done():
-			log.Log.Debug("Write proc: ctx cancelled req")
+			log.Debug("Write proc: ctx cancelled req")
 			return ctx.Err()
 		default:
 		}
 		if !w.retryQueue.isEmpty() {
-			log.Log.Debug("Write proc: taking batch from retry queue")
+			log.Debug("Write proc: taking batch from retry queue")
 			if !retrying {
 				b := w.retryQueue.first()
 				// Can we write? In case of retryable error we must wait a bit
 				if w.lastWriteAttempt.IsZero() || time.Now().After(w.lastWriteAttempt.Add(time.Millisecond*time.Duration(b.retryInterval))) {
 					retrying = true
 				} else {
-					log.Log.Warn("Write proc: cannot write yet, storing batch to queue")
+					log.Warn("Write proc: cannot write yet, storing batch to queue")
 					w.retryQueue.push(batch)
 					batchToWrite = nil
 				}
@@ -85,7 +85,7 @@ func (w *Service) HandleWrite(ctx context.Context, batch *Batch) error {
 				batchToWrite.retries++
 				if batch != nil {
 					if w.retryQueue.push(batch) {
-						log.Log.Warn("Write proc: Retry buffer full, discarding oldest batch")
+						log.Warn("Write proc: Retry buffer full, discarding oldest batch")
 					}
 					batch = nil
 				}
@@ -107,12 +107,12 @@ func (w *Service) HandleWrite(ctx context.Context, batch *Batch) error {
 func (w *Service) WriteBatch(ctx context.Context, batch *Batch) error {
 	wURL, err := w.WriteURL()
 	if err != nil {
-		log.Log.Errorf("%s\n", err.Error())
+		log.Errorf("%s\n", err.Error())
 		return err
 	}
 	var body io.Reader
 	body = strings.NewReader(batch.batch)
-	log.Log.Debugf("Writing batch: %s", batch.batch)
+	log.Debugf("Writing batch: %s", batch.batch)
 	if w.writeOptions.UseGZip() {
 		body, err = gzip.CompressWithGzip(body)
 		if err != nil {
@@ -133,7 +133,7 @@ func (w *Service) WriteBatch(ctx context.Context, batch *Batch) error {
 
 	if perror != nil {
 		if perror.StatusCode == http.StatusTooManyRequests || perror.StatusCode == http.StatusServiceUnavailable {
-			log.Log.Errorf("Write error: %s\nBatch kept for retrying\n", perror.Error())
+			log.Errorf("Write error: %s\nBatch kept for retrying\n", perror.Error())
 			if perror.RetryAfter > 0 {
 				batch.retryInterval = perror.RetryAfter * 1000
 			} else {
@@ -141,11 +141,11 @@ func (w *Service) WriteBatch(ctx context.Context, batch *Batch) error {
 			}
 			if batch.retries < w.writeOptions.MaxRetries() {
 				if w.retryQueue.push(batch) {
-					log.Log.Warn("Retry buffer full, discarding oldest batch")
+					log.Warn("Retry buffer full, discarding oldest batch")
 				}
 			}
 		} else {
-			log.Log.Errorf("Write error: %s\n", perror.Error())
+			log.Errorf("Write error: %s\n", perror.Error())
 		}
 		return perror
 	}